feat: refactor sanic benchmark test util into fixtures
