+++ tests/test_request_stream.py
@@ -279,9 +279,12 @@ def test_request_stream_blueprint(app):
                 if body is None:
                     break
                 await response.write(body.decode("utf-8"))
+
         return stream(streaming)
 
-    bp.add_route(post_add_route, '/post/add_route', methods=['POST'], stream=True)
+    bp.add_route(
+        post_add_route, "/post/add_route", methods=["POST"], stream=True
+    )
     app.blueprint(bp)
 
     assert app.is_request_stream is True
